+++ src/components/pages/WelcomePage.jsx
@@ -1,10 +1,35 @@
 import styled from "styled-components";
 import Input from "../UI/atoms/Input";
 import Button from "../UI/atoms/Button";
-import WelcomePageHook from "../functions/WelcomePage";
+import PropTypes from "prop-types";
 
-const WelcomePage = ({ onWelcomePageSubmit }) => {
-  const { username, onInputChange, onInputKeyDown } = WelcomePageHook();
+const WelcomePage = ({
+  username,
+  setUsername,
+  tasks,
+  toggleShowWelcomeMessage,
+  setFirstLoad,
+}) => {
+  const onWelcomePageSubmit = (username) => {
+    setUsername(username);
+    localStorage.setItem("tasks", JSON.stringify(tasks));
+    localStorage.setItem("username", username);
+    setFirstLoad(false);
+    setTimeout(() => {
+      toggleShowWelcomeMessage(false);
+    }, 10000);
+  };
+
+  const onInputChange = (e) => {
+    setUsername(
+      e.target.value.charAt(0).toUpperCase() +
+        e.target.value.slice(1).toLowerCase()
+    );
+  };
+
+  const onInputKeyDown = (e) => {
+    if (isFinite(+e.key)) e.preventDefault();
+  };
 
   return (
     <StyledWelcomePage>
@@ -18,8 +43,8 @@ const WelcomePage = ({ onWelcomePageSubmit }) => {
         alignSelf={"flex-end"}
         type={"submit"}
         onClick={(e) => {
-          e.preventDefault();
           onWelcomePageSubmit(username);
+          e.preventDefault();
         }}
       >
         Next
@@ -42,3 +67,11 @@ const StyledWelcomePage = styled.form`
     0 13px 10px rgba(0, 0, 0, 0.06), 0 22px 18px rgba(0, 0, 0, 0.072),
     0 42px 33px rgba(0, 0, 0, 0.086), 0 100px 80px rgba(0, 0, 0, 0.12);
 `;
+
+WelcomePage.propTypes = {
+  username: PropTypes.string,
+  setUsername: PropTypes.func,
+  tasks: PropTypes.array,
+  toggleShowWelcomeMessage: PropTypes.func,
+  setFirstLoad: PropTypes.func,
+};
